server/pyproject.toml
[project]
name = "bossypaints"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloud-volume>=11.0.2",
    "fastapi>=0.115.4",
    "httpx>=0.27.2",
    "intern>=1.4.1",
    "jque>=0.1.3",
    "numpy>=2.1.3",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "scikit-image>=0.24.0",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.7.3",
]
